PetShopDjango/requirements.txt
django
pillow
django-admin-interface
django-crispy-forms
django-allauth
django-rest-framework
django-pwa